requirements.txt
requests==2.20.1
dropbox==9.2.0
Flask==1.0.2
psycopg2==2.7.6
google-api-python-client==1.7.6
python-dotenv==0.10.0
oauth2client==4.1.3
